+++ packages/core/src/Beemo.ts
@@ -204,12 +204,17 @@ export default class Beemo {
     const driver = tool.getPlugin('driver', driverName);
     const context = this.prepareContext(new DriverContext(args, driver, parallelArgv));
 
-    tool.emit(`${driverName}.init-driver`, [context, driver]);
+    tool.emit(`${context.eventName}.init-driver`, [context, driver]);
     tool.debug('Running with %s driver', driverName);
 
     return this.startPipeline(context)
       .pipe(new ConfigureRoutine('config', tool.msg('app:configGenerate')))
-      .pipe(new ExecuteDriverRoutine('driver', tool.msg('app:driverExecute')))
+      .pipe(
+        new ExecuteDriverRoutine(
+          'driver',
+          tool.msg('app:driverExecute', { name: driver.metadata.title }),
+        ),
+      )
       .pipe(new CleanupRoutine('cleanup', tool.msg('app:cleanup')))
       .run(driverName);
   }
@@ -221,12 +226,17 @@ export default class Beemo {
     const { tool } = this;
     const context = this.prepareContext(new ScriptContext(args, scriptName));
 
-    tool.emit(`${scriptName}.init-script`, [context, scriptName]);
-    tool.debug('Running with %s script', scriptName);
+    tool.emit(`${context.eventName}.init-script`, [context, scriptName]);
+    tool.debug('Running with %s script', context.scriptName);
 
     return this.startPipeline(context)
-      .pipe(new ExecuteScriptRoutine('script', tool.msg('app:scriptExecute', { name: scriptName })))
-      .run(scriptName);
+      .pipe(
+        new ExecuteScriptRoutine(
+          'script',
+          tool.msg('app:scriptExecute', { name: context.scriptName }),
+        ),
+      )
+      .run();
   }
 
   /**